pyproject.toml
[project]
name = "tg-chatgpt-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "k1tchenboyPS",email = "pozdniakoff.alex@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot (>=22.1,<23.0)",
    "openai (>=1.86.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "openai-whisper @ git+https://github.com/openai/whisper.git",
    "gtts (>=2.5.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
